backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
opencv-python==4.8.1.78
numpy==1.24.3
onnxruntime==1.15.1
scikit-learn==1.3.0
Pillow==10.0.0
insightface==0.7.3
mysql-connector-python==8.1.0
